requirements.txt
langchain
openai
Pillow
stability-sdk
streamlit
streamlit-extras
google-genai
google-api-core
groq